requirements.txt
streamlit
langchain
langchain-community
langchain-core
numpy
altair
transformers
sentence-transformers
python-dotenv
faiss-cpu
chromadb
torch==1.12.0
pdfminer.six
pillow_heif
opencv-python
unstructured
langc
python-dotenv
pdf2image
pytesseract
unstructured_inference
